% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_chemoprot_labeled.R
\name{clean_chemoprot_labeled}
\alias{clean_chemoprot_labeled}
\title{Clean Labeled Chemoproteomics Data}
\usage{
clean_chemoprot_labeled(
  dat_list,
  tab_names,
  mage_cols,
  analysis_cols,
  reporter_cols,
  fdata_cols,
  protein_collection_cols
)
}
\arguments{
\item{dat_list}{List containing information from each of the different TMT excel sheets}

\item{tab_names}{1 x n data.frame specifying the names of the n different excel sheets
in the TMT excel file and what function to which they correspond. At the time the function
was written, n = 9}

\item{mage_cols}{1 x n data.frame specifying the n mage column names that are used
for analyses. At the time the function was written, n = 6}

\item{analysis_cols}{1 x n data.frame specifying the n package analysis column names
that are used for analyses. At the time the function was written, n = 2}

\item{reporter_cols}{1 x n data.frame specifying the n reporter ions typed column names
that are used for analyses. At the time the function was written, n = 2}

\item{fdata_cols}{1 x n data.frame specifying the n f_data column names that are used
for analyses. At the time the function was written, n = 6}

\item{protein_collection_cols}{1 x n data.frame specifying the n protein collection
column names that are used for analyses. At the time the function was written, n = 1}
}
\value{
List of isobaricpepObjects that have been cleaned and ready for further labeled chemoproteomics analysis
}
\description{
This function takes labeled chemoproteomics data from an excel format
and returns an S3 object compatible with another R package, pmartR, that
can then be ran in a markdown file for further filtering, normalization,
and statistical analyses
}
\examples{
library(chemoprotR)
data("label_pmart")
tb_name <- data.frame(analysis_jobs = "T_Data_Package_Analysis_Jobs",
                        reporter_ions = "T_Reporter_Ions_Typed",
                        mage = "Mage",
                        protein_collection = "Protein_collection_data",
                        fdata = "f_data")
mg_col <- data.frame(job_name = "Job",
                        scan_name = "Scan",
                        peptide_name = "Peptide",
                        protein_name = "Protein",
                        qvalue_name = "QValue",
                        msgf_specprob_name = "MSGF_SpecProb")
anlys_col <- data.frame(job_name = "Job",
                            dataset_id_name = "Dataset_ID")
rep_col <- data.frame(dataset_id_name = "Dataset",
                            scan_name = "ScanNumber")
f_col <- data.frame(sampleID_name = "SampleID",
                         group_name = "Grouping",
                         job_name = "Job",
                         replicate_name = "Replicate",
                         plex_name = "Plex",
                         ionization_name = "Ionization")
pro_col <- data.frame(proteinName_name = "protein_name")
htp_clean <- clean_chemoprot_labeled(label_setup,tb_name,mg_col,anlys_col,rep_col,f_col,pro_col)

}
\author{
Damon Leach, Kelly Stratton
}
