% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_chemoprot_unlabeled.R
\name{clean_chemoprot_unlabeled}
\alias{clean_chemoprot_unlabeled}
\title{Clean Unlabeled Chemoproteomics Data}
\usage{
clean_chemoprot_unlabeled(
  dat_list,
  tab_names,
  mage_cols,
  fdata_cols,
  protein_collection_cols
)
}
\arguments{
\item{dat_list}{List containing information from each of the different TMT excel sheets}

\item{tab_names}{1 x n data.frame specifying the names of the n different excel sheets
in the TMT excel file required for analysis. At the time the function was written,
n = 4}

\item{mage_cols}{1 x n data.frame specifying the n mage column names that are used
for analyses. At the time that the function was written, n = 9}

\item{fdata_cols}{1 x n data.frame specifying the n f_data column names that are used
for analyses. At the time the function was written, n = 9}

\item{protein_collection_cols}{1 x n data.frame specifying the protein collection
column names that are used for analyses. At the time the function was written, n = 1}
}
\value{
List of pepObjects that have been cleaned and ready for further unlabeled chemoproteomics analysis
}
\description{
This function takes unlabeled chemoproteomics data from an excel format
and returns an S3 object compatible with another R package, pmartR, that
can then be ran in a markdown file for further filtering, normalization,
and statistical analyses
}
\examples{
library(chemoprotR)
data("labelfree_setup")
tb_name <- data.frame(mage = "Mage",
                        metadata = "Metadata",
                        protein_collection = "Protein collection list",
                        fdata = "f_data")
mg_col <- data.frame(job_name = "Job",
                        scan_name = "Scan",
                        peptide_name = "Peptide",
                        protein_name = "Protein",
                        qvalue_name = "QValue",
                        pepqvalue_name = "PepQValue",
                        total_ion_intensity_name = "TotalIonIntensity",
                        peak_area_name = "PeakArea",
                        msgf_specprob_name = "MSGF_SpecProb")
f_col <- data.frame(sampleID_name = "SampleID",
                         group_name = "Grouping",
                         rep_name = "biological_replicate",
                         job_name = "Job")
pro_col <- data.frame(proteinName_name = "protein_name",
                      proteinCollectionID_name = "protein_collection_id",
                      proteinCollection_name = "protein_collection",
                      description_name = "description",
                      referenceID_name = "reference_id",
                      residueCount_name = "residue_count",
                      monotopicMass_name = "monoisotopic_mass",
                      proteinID_name = "protein_id")
normalization_info <- data.frame(norm_fn = "mean",
                                 backtransform = TRUE)
htp_pmart_cleaned <- clean_chemoprot_unlabeled(labelfree_setup,tb_name,mg_col,f_col,pro_col)

}
\author{
Damon Leach, Kelly Stratton
}
