% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{htp_normalize}
\alias{htp_normalize}
\title{High-Throughput (HTP) Normalization}
\usage{
htp_normalize(pmartObj, norm_function = "mean", backtransform_val = TRUE)
}
\arguments{
\item{pmartObj}{S3 omicsData object of the class 'pepData'}

\item{norm_function}{character string specifying what normalization method is
to be used (either mean or median)}

\item{backtransform_val}{logical indicator for whether the data should be backtransformed
after normalization has been applied}
}
\value{
an object of the class "pepData" that has undergone high-throughput normalization
}
\description{
This function normalizes high-throughput peptide-level data
}
\examples{
library(pmartR) 
library(chemoprotR)
data(label_pmart)
pep_object <- pmartR::edata_transform(label_pmart,"log2")
pep_isobaric <- normalize_isobaric(pep_object,exp_cname = "Plex",apply_norm = TRUE,
    channel_cname = "Grouping",refpool_channel = "Reference")
pep_group <- pmartR::group_designation(pep_isobaric,main_effects = "Grouping")
pep_norm <- htp_normalize(pep_group,"mean",TRUE)

}
\author{
Damon Leach, Kelly Stratton
}
