% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{label_pmart}
\alias{label_pmart}
\title{Cleaned Labeled HTP pmartR-friendly Object}
\format{
An  isobaricpepData object (see
\code{\link[pmartR]{as.isobaricpepData}} for details)
\describe{
\item{e_data}{a \eqn{12665 \times 12} data frame of expression data.
Each row corresponds to expression data for each isobarically labeled peptide}
\item{f_data}{a data frame with \eqn{1 \times 6} data frame of feature data.
Each row corresponds to a sample with a column giving the unique sample identifiers found in e_data
column names, the grouping, job number, replicate, plex, and ionization}
\item{e_meta}{a data frame with \eqn{12665 \times 14} data frame of biomolecule information.
Each row corresponds to data pertaining to biomolecule information for each isobarically labeled peptide}
}
}
\description{
These data originates from the label_setup dataset and pertain to values from Job 2160486. However, the data
has been transformed and filtered such that the data is compatible with pmartR
}
