% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{label_setup}
\alias{label_setup}
\title{Labeled HTP Chemoproteomics Setup File}
\format{
A List object of Length 7
\describe{
\item{T_alias}{a data frame with dimensions \eqn{22 \times 3} pertaining to alias data}
\item{T_data}{a data frame with dimensions \eqn{2890 \times 22} pertaining to peptide-level data}
\item{T_Data_Package_Analysis_Jobs}{a data frame with dimensions \eqn{4 \times 25} pertaining to T data package analysis jobs data
where each row corresponds to a different job}
\item{T_Reporter_Ions_Typed}{a data frame with dimensions \eqn{28188 \times 19} data pertaining to the reported intensity for each of the
ionizations}
\item{Mage}{a data frame with dimensions \eqn{39} pertaining to Mage information}
\item{f_data}{a data frame with dimensions \eqn{22 \times 6} pertaining to feature data.
Each row corresponds to a sample with a column giving the unique sample identifiers found in e_data
column names, the injection order, the group information, and the batch information.}
\item{Protein_collection_data}{a data frame with dimensions \eqn{5564 \times 8} pertaining to protein collection data}
}
}
\description{
These data are the original formatting of the high-throughput (HTP) chemoproteomics
file after loading in the data from excel.
}
