% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{labelfree_pmart}
\alias{labelfree_pmart}
\title{Cleaned Labeled HTP pmartR-friendly Object}
\format{
A pepData object (see
\code{\link[pmartR]{as.pepData}} for details)
\describe{
\item{e_data}{a \eqn{294655 \times 10} data frame of expression data.
Each row corresponds to expression data for each label-free peptide}
\item{f_data}{a data frame with \eqn{9 \times 5} data frame of feature data.
Each row corresponds to a sample with a column giving the unique sample identifiers found in e_data
column names, the grouping, job, replicate, and job number}
\item{e_meta}{a \eqn{294655 \times 16} data frame of expression data.
Each row corresponds to data pertaining to biomolecule information for each label-free peptide}
}
}
\description{
These data originates from the labelfree_setup dataset. However, the data
has been transformed and filtered such that the data is compatible with pmartR
}
