% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{labelfree_setup}
\alias{labelfree_setup}
\title{Label-Free HTP Chemoproteomics Setup File}
\format{
A List object of Length 4
\describe{
\item{Metadata}{a data frame with dimensions \eqn{6 \times 27} pertaining to metadata where each row
corresponds to a different job}
\item{f_data}{a data frame with \eqn{6 \times 4} data frame of feature data. Each row corresponds to a sample with a column giving the unique sample identifiers found in e_data
column names, the injection order, the group information, and the batch information.}
\item{Mage}{a data frame with dimensions \eqn{132732 \times 39} pertaining to mage information, and stores
the Total Ion Intensity values}
\item{Protein collection list}{a data frame with dimensions \eqn{243824 \times 8} pertaining to protein collection data}
}
}
\description{
These data are the original formatting of the high-throughput (HTP) chemoproteomics
file after loading in the data from excel.
}
