% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundancy_sum.R
\name{redundancy_sum}
\alias{redundancy_sum}
\title{Peptide Redundancy Summation}
\usage{
redundancy_sum(pmartObj, mage_names, protein_collect_names, labeled = TRUE)
}
\arguments{
\item{pmartObj}{S3 omicsData object of the class 'pepData'}

\item{mage_names}{1 x n data.frame specifying the n mage column names that are used
for analyses}

\item{protein_collect_names}{1 x n data.frame specifying the n protein collection
column names that are used for analyses}

\item{labeled}{logical argument specifying whether the peptide level data
is labeled or unlabeled (TRUE indicating labeled and FALSE indicating unlabeled)}
}
\value{
S3 object that has been summed to remove redundancies
}
\description{
This function takes an S3 object of class pepData that contains either
labeled or unlabeled chemoproteomics data and sums up peptides to remove
redundancies such that each row in the e_data corresponds to a single unique
peptide
}
\examples{
library(chemoprotR)
data(label_pmart)
prot_collect <- data.frame(proteinName_name = "protein_name")
mage_info <- data.frame(job_name = "Job",
                        scan_name = "Scan",
                        peptide_name = "Peptide",
                        protein_name = "Protein",
                        qvalue_name = "QValue",
                        pepqvalue_name = "PepQValue",
                        total_ion_intensity_name = "TotalIonIntensity",
                        peak_area_name = "PeakArea",
                        msgf_specprob_name = "MSGF_SpecProb")
htp_no_redundancy <- redundancy_sum(pmartObj = label_pmart,mage_names = mage_info,
                                    protein_collect_names = prot_collect,labeled = TRUE)
                                    
}
\author{
Damon Leach, Kelly Stratton
}
