% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{median}
\alias{setNames}
\alias{i_am}
\alias{read_excel}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{here}{\code{\link[here]{i_am}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{readxl}{\code{\link[readxl]{read_excel}}}

  \item{stats}{\code{\link[stats]{median}}, \code{\link[stats]{setNames}}}
}}

