% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summation_rollup.R
\name{sum_pep_function}
\alias{sum_pep_function}
\title{Sum Peptide Helper Function}
\usage{
sum_pep_function(pepData, top3)
}
\arguments{
\item{pepData}{an omicsData object of the class 'pepData'}

\item{top3}{logical indicating whether to use top 3 most abundant values or all values}
}
\value{
list with information regarding edata and emeta
}
\description{
This function aids in the process of summing up peptide abundance values
for rollup via summation
}
\author{
Rachel Richardson, Damon Leach, Kelly Stratton
}
