% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summation_rollup.R
\name{sum_pep_quant}
\alias{sum_pep_quant}
\title{Sum Peptide Quantitation}
\usage{
sum_pep_quant(pepData, top3 = F)
}
\arguments{
\item{pepData}{an omicsData object of the class 'pepData'}

\item{top3}{logical indicating whether to use top 3 most abundant values or all values (default is FALSE)}
}
\value{
an omicsData object of the class 'proData'
}
\description{
This function sums up peptide abundance value for rollup
}
\author{
Rachel Richardson, Damon Leach, Kelly Stratton
}
